KAFKA-13114; Revert state and reregister raft listener (#11116)
RaftClient's scheduleAppend may split the list of records into multiple
batches. This means that it is possible for the active controller to
see a committed offset for which it doesn't have an in-memory snapshot.

If the active controller needs to renounce and it is missing an
in-memory snapshot, then revert the state and reregister the Raft
listener. This will cause the controller to replay the entire metadata
partition.

Reviewers: Jason Gustafson <jason@confluent.io>

+++ metadata/src/main/java/org/apache/kafka/controller/Controller.java
@@ -263,7 +263,7 @@ public interface Controller extends AutoCloseable {
      * If this controller is active, this is the non-negative controller epoch.
      * Otherwise, this is -1.
      */
-    long curClaimEpoch();
+    int curClaimEpoch();
 
     /**
      * Returns true if this controller is currently active.
